sumodata: search for retired rikishi during basho import
When searching for rikishi during the import of a basho, it's possible
that the rikishi referenced in matchup data is retired. When pulling
rikishi data from the SumoAPI, we might need to search for retired
rikishi to find the data.

This patch adds the fallback path to search for retired rikishi

Signed-off-by: Jeremy C. Andrus <jchristian.andrus@gmail.com>